Rename test resource directories
