+++ elasticsearch-service/test/elasticsearchBuildersTest/elasticsearchBuilderTestResources/indexPatternBuilder.spec.resources.ts
